Add test cases for complex type validation. Refs #392
